feat: ResetStates
